Add files via upload
"Nuclear.mo" is the first-of-kind nuclear package for Modelica 3.2.1 with a SingleFluidMoltenSalt reactor model and a DualFluidMoltenSaltReactor model. The "jModScript" txt-file contains some lines to run these models with the help of JModelica. For the time being the numerical solver of OpenModelica is not able to deal with the complexity of these two models.